// src/interface/personal-timekeeping/type.ts
export interface IGetAttendanceParams {
  params: IGetAttendanceContentParams;
}

export interface IGetAttendanceContentParams {
  args: any[];
}
// ...
export interface WeeklyReportState {
  id?: number;
  date?: any;
  employee_code?: string;
  total_work_time?: number;
  shift_start?: number;
  shift_end?: number;
  total_shift_work_time?: number;
  holiday_start?: number;
  holiday_end?: number;
  rest_start?: number;
  rest_end?: number;
  employee_name?: string;
  shift_name?: string;
  explanation?: boolean;
  is_explanation?: boolean;
  type_shifts?: number;
  on_time?: boolean;
  wrong_time?: boolean;
  rest_shifts?: boolean;
  empty_shifts?: boolean;
  extra_hour?: number;
  approval_hour?: number;
  company?: string;
  department?: string;
  job_title?: any;
  surplus_minute_work?: number;
  attendance_attempt_1?: string;
  attendance_attempt_2?: string;
  attendance_attempt_3?: string;
  attendance_attempt_4?: string;
  attendance_attempt_5?: string;
  attendance_attempt_6?: string;
  attendance_attempt_7?: string;
  attendance_attempt_8?: string;
  attendance_attempt_9?: string;
  attendance_attempt_10?: string;
  attendance_attempt_11?: string;
  attendance_attempt_12?: string;
  attendance_attempt_13?: string;
  attendance_attempt_14?: string;
  attendance_attempt_15?: string;
  last_attendance_attempt?: string;
  attendance_inout_1?: string;
  attendance_inout_2?: string;
  attendance_inout_3?: string;
  attendance_inout_4?: string;
  attendance_inout_5?: string;
  attendance_inout_6?: string;
  attendance_inout_7?: string;
  attendance_inout_8?: string;
  attendance_inout_9?: string;
  attendance_inout_10?: string;
  attendance_inout_11?: string;
  attendance_inout_12?: string;
  attendance_inout_13?: string;
  attendance_inout_14?: string;
  attendance_inout_15?: string;
  attendance_inout_last?: string;
  time_keeping_code?: string;
  resource_calendar?: number;
  missing_checkin_break?: boolean;
  split_shift?: boolean;
  leave_early?: number;
  attendance_late?: number;
  actual_total_work_time?: number;
  minutes_working_reduced?: number;
  locked:any
}